[cpp] Removed normalizing non-classical LM.
It's a shame it was added in the first place and I should never show it
anyone. The approach is blatantly incorrect and produces close-to-true
values but not really the true ones.

The root cause is that the value of the cost function is normalized by sum
of σ², but the value of the gradient of the cost function is not. What's
worse, the normalization term is different at each point on each iteration
since it depends on the values of x⃗, thus it cannot be considered to be
merely changing the gradient descent rate.

Let's throw this away and just use the wrapped models everywhere.